PYIC-8593: Refactor client side code to separate rendering from api calls.
